Re-structure component NavOptions -> NavCards
