+++ README.md
@@ -15,3 +15,5 @@
 ## Test
 
 ## TODO
+
+- [ ] Add note for `Elevation API` of Android shadow in Notion view me at `Google`
