+++ app/screens/HomeScreen.js
@@ -1,23 +1,23 @@
 import { View, Image, SafeAreaView } from 'react-native'
 
 import tw from '../../lib/tailwind'
-import NavOptions from '../components/NavOptions'
+import NavCards from '../components/NavCards'
 
 const HomeScreen = () => {
   return (
     <SafeAreaView style={tw`bg-white h-full`}>
-      <View style={tw`p-5`}>
+      <View style={tw`p-7 h-full`}>
         <Image
-          style={tw.style('ml-2', {
+          style={{
             width: 100,
             height: 100,
             resizeMode: 'contain',
-          })}
+          }}
           source={{
             uri: 'https://links.papareact.com/gzs',
           }}
         />
-        <NavOptions />
+        <NavCards />
       </View>
     </SafeAreaView>
   )
